Version 0.7.0. API changes, read on...
The previous implementation of the Weatherzone::Connection class was not thread
safe. The combination of using the singleton implementation and storing an
instance of the singleton in a class variable on the finder object resulted in
bleed-through of account details across 'perceived' multiple connections.

The API for setting up a connection changes thus:

  The old way:

    Weatherzone::Connection.connect("username", "password", :url => "url") do
      "sekret" + password
    end

  The new way:

    keygen = lambda { "sekret" + password }
    @connection = Weatherzone::Connection.new("username", "password", keygen, :url => "url")

Note that we now have an instance of a connection. We need to pass this as
the first parameter to any finder.

  The old way of finding:

    Weather.find_by_location_name("Sydney, NSW")

  The new way of finding:

    Weather.find_by_location_name(@connection, "Sydney, NSW")